Update georaster-layer-for-leaflet.ts
Changes enable low resolution image to become upscaled, and other stuff that was modified for working out how to implement this. 

Added extra square boxes to check upscaling overscan pixels as resampling will make sharp lines without a pixel outside of bounds to blend to. Changed the draw boxes for these to draw every 1/3 tiles as it was hard to see when they overlapped.